+++ packages/flame/CHANGELOG.md
@@ -17,6 +17,8 @@
  - Create default implementations of `RectangleComponent`, `CircleComponent` and `PolygonComponent`
  - Streamlined the argument list for all components extending `PositionComponent`
  - Improved interaction between viewport and isHud components
+ - Calling super-method in `.render()` is now optional
+ - Components that manipulate canvas state are now responsible for saving/restoring that state
 
 ## [1.0.0-releasecandidate.16]
  - `changePriority` no longer breaks game loop iteration